Add machine_type fact
This commit will query Apple with the model code from the machine's
serial number to determine the machine type (which comes from a
block of XML returned from the open-uri request). This process
DOES take some time (since it's hitting the web), so we probably
should wrap this and rescue any errors (should Apple change the
way they deliver this data).